Refactor Dashboard and Home components, add PlanTrip page
- Refactor Dashboard and Home components to remove unnecessary comments and imports
- Add PlanTrip page to allow users to plan their trip to Tokyo
- Update package.json to include new dependencies for Google Maps API and Axios
- Lazy load components in App.js for better performance
