+++ frontend/src/pages/Destination.jsx
@@ -1,35 +1,43 @@
-import React, { useState } from 'react'
-import { faMoneyBill1Wave, faWalking } from '@fortawesome/free-solid-svg-icons'
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import React, { useState, useEffect } from 'react';
+import { GoogleMap, LoadScript, Marker } from '@react-google-maps/api';
 import Navbar from '../components/Navbar';
-import "../styles/destination.scss"
-import Map, { Marker } from 'react-map-gl';
-import 'mapbox-gl/dist/mapbox-gl.css';
+import "../styles/destination.scss";
 
 const Destination = () => {
-
-    const [viewState, setViewState] = React.useState({
-        latitude: 35.6764,
-        longitude: 139.7300,
+    const [viewState, setViewState] = useState({
+        latitude: 35.6764,  // Default latitude (Tokyo, Japan)
+        longitude: 139.7300, // Default longitude (Tokyo, Japan)
         zoom: 10
     });
 
-    const [accommodationPrice, setAccommodationPrice] = useState(0);
-    const [flightDeparturePrice, setFlightDeparturePrice] = useState(0);
-    const [flightReturnPrice, setFlightReturnPrice] = useState(0);
-    const [transportPrice, setTransportPrice] = useState(0);
-    const [mealsPrice, setMealsPrice] = useState(0);
-    const [activitiesPrice, setActivitiesPrice] = useState(0);
+    const [error, setError] = useState(null);
 
-    const calculateTotalBudget = () => {
-        return (
-            parseFloat(accommodationPrice) +
-            parseFloat(flightDeparturePrice) +
-            parseFloat(flightReturnPrice) +
-            parseFloat(transportPrice) +
-            parseFloat(mealsPrice) +
-            parseFloat(activitiesPrice)
-        );
+    // Fetch destination data from backend
+    useEffect(() => {
+        // Replace with your API endpoint
+        fetch('/api/destination')
+            .then(response => response.json())
+            .then(data => {
+                setViewState({
+                    latitude: data.latitude,
+                    longitude: data.longitude,
+                    zoom: 10
+                });
+            })
+            .catch(error => {
+                console.error('Error fetching destination data:', error);
+                setError('Failed to load destination data.');
+            });
+    }, []);
+
+    const mapStyles = {
+        height: "300px",
+        width: "1500px"
+    };
+
+    const defaultCenter = {
+        lat: viewState.latitude,
+        lng: viewState.longitude
     };
 
     return (
@@ -48,139 +56,47 @@ const Destination = () => {
                             culinary scene. Tokyo is a dynamic blend of history,
                             technology, and culture.
                         </p>
-                        <p>
-                            <span>Vibrant</span>
-                            <span>Modern</span>
-                            <span>Bustling</span>
-                            <span>Electric</span>
-                            <span>Cultural</span>
-                        </p>
                         <section>
-                            <FontAwesomeIcon icon={faMoneyBill1Wave}
-                                className="icon" />
-                            <span>$600</span>
-                            <FontAwesomeIcon icon={faWalking}
-                                className="icon" />
-                            <span>2-3 days</span>
+                            {/* Add any information like travel budget and duration */}
                         </section>
                     </div>
                 </div>
                 <div className="location-map">
-                    <Map
-                        {...viewState}
-                        onMove={evt => setViewState(evt.viewState)}
-                        style={{ width: 1500, height: 300 }}
-                        mapStyle="mapbox://styles/mapbox/streets-v9"
-                        mapboxAccessToken={process.env.REACT_APP_MAPBOX_TOKEN}
-                    >
-                        <Marker className="marker"
-                            longitude={viewState.longitude}
-                            latitude={viewState.latitude} color="red" />
-                    </Map>
-                </div>
-
-
-                <div className="plan-trip-container">
-                    <h1>Start Planning Trip to Tokyo</h1>
-                    <div className="form-container">
-                        <div className="plan-input">
-                            <label>Start Date:</label>
-                            <input type="date" />
-                        </div>
-                        <div className="plan-input">
-                            <label>End Date:</label>
-                            <input type="date" />
-                        </div>
-                        <div className="plan-input">
-                            <label>No of People:</label>
-                            <input type="number" />
-                        </div>
-
-                        <div>
-                            <h3>Accommodation</h3>
-
-                            <div className="plan-input">
-                                <label>Accommodation Name:</label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Price: $</label>
-                                <input type="number" value={accommodationPrice}
-                                    onChange={(e) => setAccommodationPrice(e.target.value)} />
-                            </div>
-                        </div>
-                        <div>
-                            <h3>Flights</h3>
-
-                            <div className="plan-input">
-                                <label htmlFor>Enter Flight Name: </label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Flight Departure Price: $</label>
-                                <input type="number" value={flightDeparturePrice}
-                                    onChange={(e) => setFlightDeparturePrice(e.target.value)} />
-                            </div>
-                            <div className="plan-input">
-                                <label htmlFor>Enter Flight Name: </label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Flight Return Price: $</label>
-                                <input type="number" value={flightReturnPrice}
-                                    onChange={(e) => setFlightReturnPrice(e.target.value)} />
-                            </div>
-                        </div>
-                        <div>
-                            <h3>Transportation</h3>
-
-                            <div className="plan-input">
-                                <label>Transport Type</label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Transport Cost: $</label>
-                                <input type="number" value={transportPrice}
-                                    onChange={(e) => setTransportPrice(e.target.value)} />
-                            </div>
-                        </div>
-                        <div>
-                            <h3>Meals & Food</h3>
-
-                            <div className="plan-input">
-                                <label htmlFor>Enter Meals to Try: </label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Meals & Snacks Price: $</label>
-                                <input type="number" value={mealsPrice}
-                                    onChange={(e) => setMealsPrice(e.target.value)} />
-                            </div>
-                        </div>
-                        <div>
-                            <h3>Activities & Tourist Spots</h3>
-                            <div className="plan-input">
-                                <label htmlFor>Enter Activities to Try: </label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label htmlFor>Enter Tourist Spots to Explore: </label>
-                                <input type="text" />
-                            </div>
-                            <div className="plan-input">
-                                <label>Activities & Tourist Spots Price: $</label>
-                                <input type="number" value={activitiesPrice}
-                                    onChange={(e) => setActivitiesPrice(e.target.value)} />
-                            </div>
-                        </div>
-                        <div>
-                            <h2>Total Budget: {calculateTotalBudget()}</h2>
-                        </div>
-                    </div>
+                    {error ? (
+                        <div>Error: {error}</div>
+                    ) : (
+                        <LoadScript googleMapsApiKey={process.env.REACT_APP_GOOGLE_MAPS_API_KEY}>
+                            <GoogleMap
+                                mapContainerStyle={mapStyles}
+                                zoom={viewState.zoom}
+                                center={defaultCenter}
+                                onClick={(e) => {
+                                    // Handle map click to change view or add markers
+                                    setViewState({
+                                        latitude: e.latLng.lat(),
+                                        longitude: e.latLng.lng(),
+                                        zoom: viewState.zoom
+                                    });
+                                }}
+                            >
+                                <Marker
+                                    position={defaultCenter}
+                                    draggable={true}
+                                    onDragEnd={(e) => {
+                                        setViewState({
+                                            latitude: e.latLng.lat(),
+                                            longitude: e.latLng.lng(),
+                                            zoom: viewState.zoom
+                                        });
+                                    }}
+                                />
+                            </GoogleMap>
+                        </LoadScript>
+                    )}
                 </div>
             </div>
         </div>
-    )
+    );
 }
 
-export default Destination
+export default Destination;
