+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,3 @@
-// client/src/pages/Home.jsx
-
 import React, {
     useState,
     useEffect
@@ -29,18 +27,18 @@ const Home = () => {
     useEffect(() => {
         const getPlaces = async () => {
           const promise = await fetch(
-            `https://api.mapbox.com/geocoding/v5/mapbox.places/${value}.json?access_token=${process.env.REACT_APP_MAPBOX_TOKEN}`
+            `https://maps.googleapis.com/maps/api/place/findplacefromtext/${value}.json?access_token=${process.env.MAP_API_KEY}`
           );
           const data = await promise.json();
           setPlaces(data.features);
           console.log(data);
         };
-      
+
         if (value) {
           getPlaces();
         }
       }, [value]);
-      
+
 
     const handleClick = (query) => {
         setValue(query.place_name)
